Try to fix static analysis

+++ oneflow/core/auto_parallel/sbp_constructor.cpp
@@ -455,7 +455,7 @@ Maybe<HashMap<const OpNode*, HashSet<std::string>>> SbpConstructor::GetMutableOp
 void SbpConstructor::InitAvailableMemory() {
   size_t free = 0;
   size_t total = 0;
-  CudaCurrentDeviceGuard guard(GlobalProcessCtx::Rank());
+  CudaCurrentDeviceGuard guard((int32_t)(GlobalProcessCtx::Rank()));
   OF_CUDA_CHECK(cudaMemGetInfo(&free, &total));
   // The estimated memory differs from the lower bound of the peak memory by the first ratio.
   // The first ratio varies from -3% to 3.2% if not enabling nccl_use_compute_stream.
